Beta Version of BMS
===== MASTER COMMIT ====
I present to u the Beta Version of BMS.
It was coded entirely in C/C++
Data are mining using the Concept of Double-way Linked List with the help of Classes. This Allow us to store and retrieve data efficiently.
So I hope u will enjoy...
For a point of remark, this project wouldn't be possible without the active participation of all its members. And for that I would like to say a big Thanks to them.

Co-Authored-By: Amadou Olabi Yuffon <amadou-olabi@users.noreply.github.com>
Co-Authored-By: mounkambou-iut-cse-4301 <mounkambou-iut-cse-4301@users.noreply.github.com>
Co-Authored-By: aly-abdelkader-iut-cse-4301 <aly-abdelkader-iut-cse-4301@users.noreply.github.com>
Co-Authored-By: anasse_mohamed_iut_cse4301 <anasse-mohamed-iut-cse4301@users.noreply.github.com>
Co-Authored-By: shalanyuy <shalanyuy@users.noreply.github.com>